23-AI-Agent/pyproject.toml
[project]
name = "23-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.56.0",
    "gradio>=6.2.0",
    "ipython>=9.8.0",
    "openai>=2.14.0",
    "openai-agents>=0.6.4",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
